refactor: 기존의 container 컴포넌트들을 section 으로 이름 변경
