+++ src/pages/CardListPage/CardListPage.jsx
@@ -3,7 +3,7 @@ import styles from "./CardListPage.module.scss";
 import { Link } from "react-router-dom";
 
 import Label from "../../components/Label/Label";
-import CardListContainer from "../../containers/CardListContainer/CardListContainer";
+import CardListSection from "../../sections/CardListSection/CardListSection";
 import AddBoxButton from "../../components/AddBoxButton/AddBoxButton";
 import { ANIMATION, HEADER_TEXT, PAGE_PATH } from "../../constants";
 import AppContext from "../../contexts/appContext";
@@ -20,7 +20,7 @@ const CardListPage = ({}) => {
         <Label labelText={HEADER_TEXT.OWNED_CARD} />
       </header>
       <main className={cx("card-list-page__main")}>
-        <CardListContainer cardListState={cardListState} />
+        <CardListSection cardListState={cardListState} />
         <Link to={PAGE_PATH.ADD}>
           <AddBoxButton />
         </Link>
