+++ src/sections/CardListSection/CardListSection.jsx
@@ -1,6 +1,6 @@
 import { useContext } from "react";
 import classNames from "classnames/bind";
-import styles from "./CardListContainer.module.scss";
+import styles from "./CardListSection.module.scss";
 import { getCardColor } from "../../utils/cardCompany";
 
 import Card from "../../components/Card/Card";
@@ -8,7 +8,7 @@ import AppContext from "../../contexts/appContext";
 
 const cx = classNames.bind(styles);
 
-const CardListContainer = () => {
+const CardListSection = () => {
   const { cardListState } = useContext(AppContext);
 
   const cardList = cardListState.map((card) => {
@@ -28,4 +28,4 @@ const CardListContainer = () => {
   return <div className={cx("card-list-container")}>{cardList}</div>;
 };
 
-export default CardListContainer;
+export default CardListSection;
